refactor: 替换文件夹层次结构领域为新树形结构存储模块

+++ mod-file-system/src/main/java/com/ricky/folder/command/MoveFolderResponse.java
@@ -12,6 +12,6 @@ import lombok.Value;
 public class MoveFolderResponse implements Response {
 
     Integer movedFolderCount;
-    Long movedFileCount;
+    Integer movedFileCount;
 
 }
